+++ plugins/user/loginguard/loginguard.php
@@ -7,6 +7,7 @@
 
 use Akeeba\LoginGuard\Site\Helper\Tfa;
 use FOF30\Container\Container;
+use Joomla\CMS\Factory;
 use Joomla\CMS\Plugin\CMSPlugin;
 use Joomla\CMS\User\User;
 use Joomla\Utilities\ArrayHelper;
@@ -72,7 +73,7 @@ class plgUserLoginguard extends CMSPlugin
 			return true;
 		}
 
-		$layout = JFactory::getApplication()->input->getCmd('layout', 'default');
+		$layout = Factory::getApplication()->input->getCmd('layout', 'default');
 
 		if (!$this->container->platform->isBackend() && !in_array($layout, array('edit', 'default')))
 		{
@@ -95,7 +96,7 @@ class plgUserLoginguard extends CMSPlugin
 			$id = isset($data->id) ? $data->id : null;
 		}
 
-		$user = JFactory::getUser($id);
+		$user = Factory::getUser($id);
 
 		// Make sure the loaded user is the correct one
 		if ($user->id != $id)
@@ -158,24 +159,29 @@ class plgUserLoginguard extends CMSPlugin
 		/** @var User $user */
 		$user = $options['user'];
 
-		if (!is_object($user) || !(($user instanceof JUser) || ($user instanceof User)))
+		if (!is_object($user) || ($user instanceof User))
 		{
 			return;
 		}
 
-		// Make sure this user does not already have 2SV enabled
-		if ($this->needsTFA($user, $options['responseType']))
+		/**
+		 * If the user already has 2SV enabled and we need to show the captive page we won't redirect them to the 2SV
+		 * setup page, of course.
+		 */
+		if ($this->needsCaptivePage($user, $options['responseType']))
 		{
 			return;
 		}
 
-		// Make sure the user does not have a flag to not bother him again with the 2SV setup page
-		if ($this->hasFlag($user))
+		/**
+		 * If the user has already asked us to not show him the 2SV setup page we have to honour their wish.
+		 */
+		if ($this->hasDoNotShowAgainFlag($user))
 		{
 			return;
 		}
 
-		// Get the redirection URL
+		// Get the redirection URL to the 2SV setup page or custom redirection per plugin configuration
 		$url           = JRoute::_('index.php?option=com_loginguard&task=methods.display&layout=firsttime', false);
 		$configuredUrl = $this->params->get('redirecturl', null);
 
@@ -185,7 +191,7 @@ class plgUserLoginguard extends CMSPlugin
 		}
 
 		// Prepare to redirect
-		JFactory::getSession()->set('postloginredirect', $url, 'com_loginguard');
+		Factory::getSession()->set('postloginredirect', $url, 'com_loginguard');
 	}
 
 	/**
@@ -222,7 +228,7 @@ class plgUserLoginguard extends CMSPlugin
 			return true;
 		}
 
-		$db = JFactory::getDbo();
+		$db = Factory::getDbo();
 
 		// Delete user profile records
 		$query = $db->getQuery(true)
@@ -264,42 +270,23 @@ class plgUserLoginguard extends CMSPlugin
 	 *
 	 * @return  bool
 	 */
-	private function needsTFA(User $user, $responseType = null)
+	private function needsCaptivePage(User $user, $responseType = null)
 	{
-		/**
-		 * If the login type is silent (cookie, social login / single sign-on, gmail, ldap) we will not ask for 2SV. The
-		 * login risk has already been managed by the external authentication method. For your reference, the
-		 * authentication response types are as follows:
-		 *
-		 * - Joomla: username and password login
-		 * - Cookie: "Remember Me" cookie with a secure, single use token and other safeguards for the user session
-		 * - GMail: login with GMail credentials (probably no longer works)
-		 * - LDAP: Joomla's LDAP plugin
-		 * - SocialLogin: Akeeba Social Login (login with Facebook etc)
-		 */
-		$silentResponses = array('cookie', 'gmail', 'ldap', 'sociallogin');
-
-		if (is_string($responseType) && !empty($responseType) && in_array(strtolower($responseType), $silentResponses))
-		{
-			return false;
-		}
-
-		// Get the user's TFA records
+		// Get the user's 2SV records
 		/** @var \Akeeba\LoginGuard\Site\Model\Tfa $tfaModel */
 		$tfaModel = $this->container->factory->model('Tfa')->tmpInstance();
 		$records = $tfaModel->user_id($user->id)->get(true);
 
-
-		// No TFA methods? Then we obviously don't need to display a captive login page.
+		// No 2SV methods? Then we obviously don't need to display a captive login page.
 		if ($records->count() < 1)
 		{
 			return false;
 		}
 
-		// Let's get a list of all currently active TFA methods
+		// Let's get a list of all currently active 2SV methods
 		$tfaMethods = Tfa::getTfaMethods();
 
-		// If not TFA method is active we can't really display a captive login page.
+		// If not 2SV method is active we can't really display a captive login page.
 		if (empty($tfaMethods))
 		{
 			return false;
@@ -313,7 +300,7 @@ class plgUserLoginguard extends CMSPlugin
 			$methodNames[] = $tfaMethod['name'];
 		}
 
-		// Filter the records based on currently active TFA methods
+		// Filter the records based on currently active 2SV methods
 		foreach($records as $record)
 		{
 			if (in_array($record->method, $methodNames))
@@ -323,7 +310,7 @@ class plgUserLoginguard extends CMSPlugin
 			}
 		}
 
-		// No viable TFA method found. We won't show the captive page.
+		// No viable 2SV method found. We won't show the captive page.
 		return false;
 	}
 
@@ -334,9 +321,9 @@ class plgUserLoginguard extends CMSPlugin
 	 *
 	 * @return  bool
 	 */
-	private function hasFlag(User $user)
+	private function hasDoNotShowAgainFlag(User $user)
 	{
-		$db = JFactory::getDbo();
+		$db = Factory::getDbo();
 		$query = $db->getQuery(true)
 			->select($db->qn('profile_value'))
 			->from($db->qn('#__user_profiles'))