+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 3.1.0
 
+**New**
+
+* Option to disable Two Step Verification on silent login (e.g. when Remember Me is used)
+
 **Other changes**
 
 * Dropped support for Joomla! 3.4, 3.5, 3.6, 3.7 as well as PHP 5.4 and 5.5